refactor: more consistent code style
